+++ src/ensembl/src/content/app/species-selector/components/popular-species-button/PopularSpeciesButton.test.tsx
@@ -31,10 +31,12 @@ jest.mock('connected-react-router', () => ({
   push: jest.fn(() => ({ type: 'push' }))
 }));
 
-jest.mock('src/shared/hooks/useAnalyticsService', () =>
-  jest.fn(() => ({
-    trackPopularSpeciesSelect: jest.fn()
-  }))
+jest.mock(
+  'src/content/app/species-selector/hooks/useSpeciesSelectorAnalytics',
+  () =>
+    jest.fn(() => ({
+      trackPopularSpeciesSelect: jest.fn()
+    }))
 );
 
 const handleSelectedSpecies = jest.fn();
